backend/Django/requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.43.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
mysqlclient==2.0.1
oauthlib==3.1.0
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
